refactor: extract context menu target connector
