+++ vaadin-grid-flow-parent/vaadin-grid-flow/src/main/resources/META-INF/resources/frontend/gridConnector.js
@@ -1011,7 +1011,7 @@ import { isFocusable } from '@vaadin/grid/src/vaadin-grid-active-item-mixin.js';
       };
 
       grid.addEventListener('vaadin-context-menu-before-open', tryCatchWrapper(function(e) {
-        contextMenuListener(grid.$contextMenuConnector.openEvent);
+        contextMenuListener(grid.$contextMenuTargetConnector.openEvent);
       }));
 
       grid.getContextMenuBeforeOpenDetail = tryCatchWrapper(function(event) {
