+++ vaadin-grid-flow-parent/vaadin-grid-flow/src/test/java/com/vaadin/flow/component/grid/contextmenu/GridContextMenuTest.java
@@ -120,6 +120,6 @@ public class GridContextMenuTest {
 
         Mockito.verify(registration).remove();
         Mockito.verify(element)
-                .callJsFunction("$contextMenuConnector.removeConnector");
+                .callJsFunction("$contextMenuTargetConnector.removeConnector");
     }
 }
